Split CommitContract tests per command.
